Refactored Request controller and test
