+++ app/src/main/java/com/ualberta/cs/alfred/RequestESController.java
@@ -1,0 +1,8 @@
+package com.ualberta.cs.alfred;
+
+/**
+ * Created by grandmanitou on 2016-11-15.
+ */
+
+public class RequestESController {
+}
