+++ app/src/main/java/com/ualberta/cs/alfred/fragments/PendingFragment.java
@@ -1,6 +1,5 @@
 package com.ualberta.cs.alfred.fragments;
 
-import android.content.Context;
 import android.content.Intent;
 import android.content.SharedPreferences;
 import android.os.Bundle;
@@ -17,7 +16,7 @@ import android.widget.ListView;
 import com.ualberta.cs.alfred.R;
 import com.ualberta.cs.alfred.Request;
 import com.ualberta.cs.alfred.RequestDetails;
-import com.ualberta.cs.alfred.RequestElasticSearchController;
+import com.ualberta.cs.alfred.RequestESGetController;
 import com.ualberta.cs.alfred.RequestList;
 
 import java.util.ArrayList;
@@ -63,7 +62,7 @@ public class PendingFragment extends Fragment {
 
         pendingListView = (ListView) view.findViewById(R.id.pendingListView);
 
-        RequestElasticSearchController.GetRequestTask getRequestTask = new RequestElasticSearchController.GetRequestTask();
+        RequestESGetController.GetRequestTask getRequestTask = new RequestESGetController.GetRequestTask();
         ArrayList<Request> pendingList = null;
 
         if (preferences.getString("MODE", null).contentEquals("Driver Mode")) {
@@ -94,7 +93,7 @@ public class PendingFragment extends Fragment {
     }
 
     private ArrayList<Request> getRiderPendingList() {
-        RequestElasticSearchController.GetRequestTask getRequestTask = new RequestElasticSearchController.GetRequestTask();
+        RequestESGetController.GetRequestTask getRequestTask = new RequestESGetController.GetRequestTask();
         ArrayList<Request> pendingList = null;
 
         try {
@@ -114,7 +113,7 @@ public class PendingFragment extends Fragment {
         /* The request that should be retrieved are all requests that are currently with a requested status and those that
         are pending that do not include the driver on the bidlist of the request.
          */
-        RequestElasticSearchController.GetRequestTask getRequestTask = new RequestElasticSearchController.GetRequestTask();
+        RequestESGetController.GetRequestTask getRequestTask = new RequestESGetController.GetRequestTask();
         RequestList pendingList = null;
 
         try {
